Correct interval computation for event-based sim
Use intervals for error correction (not injection) in event-based
simulation, to allow for some errors to be detected simulatenously.
Otherwise, spontaneous DUE are impossible (especially with ChipKill +
agressive retirement)

Correct the way interval comparison is done (for both scrubbing and
error correction). Previously, the simulation compared the current scrub
interval with the previous scrub interval *after* injecting an error.

This leads to wrong sequences of events.  For example, in the following
timline, the simulation would not scrub between errors 2 and 3 and
instead scrub after error 3 :

	|------*---*--|---*--------|

	(* = faults, | = scrubbing interval)